requirements.txt
# --- Web Framework ---
fastapi==0.119.0
uvicorn==0.37.0
starlette==0.48.0

# --- Data / Models ---
pydantic==2.11.10
pydantic_core==2.33.2

# --- Utilities ---
python-dotenv==1.0.1
requests==2.32.5
PyYAML==6.0.3
orjson==3.11.3

# --- Text Processing (GEO metrics) ---
regex==2025.9.18
tiktoken==0.12.0
numpy==2.3.3
scipy==1.16.3

# --- Optional but useful for fast JSON ---
typing_extensions==4.15.0
